requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.11
django-cors-headers==4.9.0
dj-database-url==2.2.0
whitenoise==6.8.2
python-decouple==3.8
gunicorn==23.0.0
boto3==1.40.51